+++ src/integTest/groovy/com/monits/gradle/sca/PmdIntegTest.groovy
@@ -18,6 +18,7 @@ import com.monits.gradle.sca.io.TestFile
 import groovy.transform.CompileDynamic
 import org.gradle.testkit.runner.BuildResult
 import org.gradle.util.GradleVersion
+import spock.lang.Ignore
 import spock.lang.Unroll
 
 import static org.gradle.testkit.runner.TaskOutcome.FAILED
@@ -308,6 +309,23 @@ class PmdIntegTest extends AbstractPerSourceSetPluginIntegTestFixture {
         result.task(':downloadPmdXmlMainPmd6').outcome == SUCCESS
         result.task(':downloadPmdXmlTestPmd6').outcome == SUCCESS
         assertThat(result.output, containsString('Running in offline mode. Using a possibly outdated version of'))
+    }
+
+    @Ignore // TODO : This test requires PMD 6.18.0 to be released
+    @SuppressWarnings('MethodName')
+    void 'incremental analysis is enabled'() {
+        given:
+        writeBuildFile()
+        goodCode()
+
+        when:
+        gradleRunner()
+            .withArguments('check', '--stacktrace')
+            .build()
+
+        then:
+        file('build/tmp/pmdMain/incremental.cache').exists()
+        file('build/tmp/pmdTest/incremental.cache').exists()
     }
 
     @Override